+++ test/src/test/node-test-driver.spec.js
@@ -1,8 +1,8 @@
 import test from 'tape-catch';
 
-import {NodeTestDriver} from 'probe.gl/test';
+import {BrowserDriver} from 'probe.gl/test';
 
 test('Top-level imports', t => {
-  t.ok(NodeTestDriver, 'NodeTestDriver symbol imported');
+  t.ok(BrowserDriver, 'BrowserDriver symbol imported');
   t.end();
 });